+++ ChangeLog.md
@@ -3,6 +3,7 @@ Web frontends change log
 
 ## ?.?.? / ????-??-??
 
+* Made HTTP response headers controllable via `View::header()` - @thekid
 * Made HTTP response status controllable via `View::status()` - @thekid
 
 ## 0.1.0 / 2018-04-02